Remove smoke-tests and move other tests to right location
